backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.1

# HTTP client for testing
httpx==0.25.2

# File processing
pypdf==3.17.4
python-docx==1.1.0
lxml==4.9.3
beautifulsoup4==4.12.2  # For EPUB and FB2 parsing
zipfile36==0.1.3  # For EPUB archives
chardet==5.2.0  # For text encoding detection

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Logging
loguru==0.7.2

# Environment
python-dotenv==1.0.0
